# tests/backend/parity/golden/openapi_schema.json
{
  "components": {
    "schemas": {
...
      "AisKeyIn": {
        "description": "Body for PUT /api/sea/ais-key.",
        "properties": {
          "key": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Key",
            "type": "string"
          }
        },
        "required": [
          "key"
        ],
        "title": "AisKeyIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SENTINEL API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/sea/ais-key": {
      "delete": {
        "description": "Forget the saved key; the ``.env`` key (if any) applies again.",
        "operationId": "delete_ais_key_api_sea_ais_key_delete",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Delete Ais Key",
        "tags": [
          "sea"
        ]
      },
      "get": {
        "description": "Report whether a key is configured and where it came from — never the key.",
        "operationId": "get_ais_key_status_api_sea_ais_key_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Ais Key Status",
        "tags": [
          "sea"
        ]
      },
      "put": {
        "description": "Save the AISStream key. Takes effect on the next watchdog tick.",
        "operationId": "put_ais_key_api_sea_ais_key_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AisKeyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Ais Key",
        "tags": [
          "sea"
        ]
      }
    },
...
    "/api/sea/vessels": {
      "get": {
        "description": "Return the current vessel picture, newest position first.\n\n``bbox`` limits the rows to the map viewport (pass it — a worldwide\nsubscription is tens of thousands of vessels). The response always carries\nthe feed status so the map can label cached vessels as STALE while the\nupstream is silent or reconnecting, rather than silently drawing old data.",
        "operationId": "get_vessels_api_sea_vessels_get",
        "parameters": [
          {
            "description": "south,west,north,east in degrees",
            "in": "query",
            "name": "bbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "south,west,north,east in degrees",
              "title": "Bbox"
            }
          },
          {
            "in": "query",
            "name": "max_rows",
            "required": false,
            "schema": {
              "default": 12000,
              "maximum": 50000,
              "minimum": 1,
              "title": "Max Rows",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Vessels",
        "tags": [
          "sea"
        ]
      }
    },
    "/api/sea/vessels/{mmsi}/track": {
      "get": {
        "description": "Return one vessel's recent path (oldest fix first) plus its current record.\n\n404 when the vessel is not in the store — it has never been heard, or it\nfell outside the retention window.",
        "operationId": "get_vessel_track_api_sea_vessels__mmsi__track_get",
        "parameters": [
          {
            "in": "path",
            "name": "mmsi",
            "required": true,
            "schema": {
              "title": "Mmsi",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Vessel Track",
        "tags": [
          "sea"
        ]
      }
    },
...
  }
}